// src/app/components/Calendar/CalendarEditScheduleDialog/types.ts
export type FormValues = {
  label: string
  startedDate: string
  startedTime: string
  finishedTime: string
  color: string
  borderColor: string
  backgroundColor: string
  errorIcon: number
  row: number
  status: string
}

export type ScheduleTheme = {
  color: string
  borderColor: string
  backgroundColor: string
  errorIcon: boolean
}